RTEMS 4.10 newlib patch 20110114
Revert "RTEMS 4.10 newlib patch 20101006"

apply: newlib-1.18.0-rtems4.10-20110114.diff

+++ newlib/libc/string/memccpy.c
@@ -35,13 +35,14 @@ PORTABILITY
 	*/
 
 #include <_ansi.h>
+#include <stdint.h>
 #include <stddef.h>
 #include <string.h>
 #include <limits.h>
 
 /* Nonzero if either X or Y is not aligned on a "long" boundary.  */
 #define UNALIGNED(X, Y) \
-  (((long)X & (sizeof (long) - 1)) | ((long)Y & (sizeof (long) - 1)))
+  (((intptr_t)X & (sizeof (long) - 1)) | ((intptr_t)Y & (sizeof (long) - 1)))
 
 /* How many bytes are copied each iteration of the word copy loop.  */
 #define LITTLEBLOCKSIZE (sizeof (long))
